requirements.txt
attrs==21.4.0
cffi==1.15.0
coverage==6.3.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.8
pytest==7.1.2
tomli==2.0.1
